feat: initial refact to use item attributes or item type attributes, some refacts on tests to support itemTypeAttributes and itemAttributes in all items creations.

+++ src/Core/NeoServer.Domain/Common/Contracts/Items/Weapons/Attributes/IHasDefense.cs
@@ -2,5 +2,5 @@ namespace NeoServer.Domain.Common.Contracts.Items.Weapons.Attributes;
 
 public interface IHasDefense
 {
-    byte Defense { get; }
+    ushort Defense { get; }
 }
